requirements.txt
django==1.4.3
south
django-allauth
django-debug-toolbar
docutils
